packages/nearestNeighbor/requirements.txt
pandas
SQLAlchemy
PyMySQL
python-dotenv
joblib
functions_framework
scikit-learn==1.4.1.post1
google-cloud-storage
sendgrid